+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Yii Framework 2 authclient extension Change Log
 
 - Bug #364: Use issuer claim from OpenID Configuration (radwouters)
 - Enh: #367: Throw more specific `ClientErrorResponseException` when the response code in `BaseOAuth::sendRequest()` is a 4xx (rhertogh)  
+- Enh #381: Add OpenID Connect client Class `yii\authclient\clients\LinkedInOIDC` for LinkedIn (mohammedNali)
 
 
 2.2.14 November 18, 2022